Normalize all messages and update messageList component to conform to new state structure
