+++ client/src/containers/MessageContainer.js
@@ -2,9 +2,11 @@ import { connect } from 'react-redux';
 import MessageList from '../components/MessageList';
 import { addMessage, fetchMessages } from '../actions/messages';
 
-const mapStateToProps = ({ messages, socket, currentChannel, profiles}) => { 
+const mapStateToProps = ({ messages, messagesByChannel, socket, currentChannel, profiles}) => { 
   return {
-    messages: messages.filter(message => message.channel_id === currentChannel.id),
+    // messages: messages.filter(message => message.channel_id === currentChannel.id),
+    messages,
+    messagesByChannel,
     socket,
     currentChannel,
     profiles
