+++ client/src/reducers/currentChannel.js
@@ -1,6 +1,6 @@
 import { SELECT_CHANNEL } from '../actions/actionTypes';
 
-const currentChannel = (state = {}, action) => {
+const currentChannel = (state = {id: 0}, action) => {
   switch (action.type) {
   case SELECT_CHANNEL:
     return action.channel;
