+++ client/src/reducers/messages.js
@@ -1,16 +1,15 @@
-import { SET_MESSAGES, ADD_MESSAGE, RECEIVE_MESSAGES, SEND_MESSAGE } from '../actions/actionTypes';
+import { SET_MESSAGES, ADD_MESSAGE } from '../actions/actionTypes';
 
-const messages = (state = [], action) => {
+const messages = (state = { }, action) => {
   switch (action.type) {
   case SET_MESSAGES:
-    return [...state, ...action.messages];
+    const messageNormal = {...state};
+    action.messages.forEach(message => {
+      messageNormal[message.id] = message;
+    });
+    return messageNormal;
   case ADD_MESSAGE:
-    action.message.created_at = Date();
-    return [...state, action.message];
-  case RECEIVE_MESSAGES:
-    //unused
-    return action.messages;
-  case SEND_MESSAGE:
+    return {...state, [action.message.id]: action.message};
   default:
     return state;
   }
